Pull code insights constants into separate file
