+++ client/web/src/enterprise/insights/core/backend/gql-backend/methods/get-backend-insight-data/deserializators.ts
@@ -5,10 +5,8 @@ import { BackendInsightData } from '../../../code-insights-backend-types'
 import { createComputeCategoricalChart } from '../../../utils/create-categorical-content'
 import { createLineChartContent } from '../../../utils/create-line-chart-content'
 
-export const MAX_NUMBER_OF_SERIES = 20
-
 export const createBackendInsightData = (insight: BackendInsight, response: InsightDataNode): BackendInsightData => {
-    const seriesData = response.dataSeries.slice(0, MAX_NUMBER_OF_SERIES)
+    const seriesData = response.dataSeries
     const isFetchingHistoricalData = seriesData.some(
         ({ status: { pendingJobs, backfillQueuedAt } }) => pendingJobs > 0 || backfillQueuedAt === null
     )
