+++ client/web/src/enterprise/insights/core/backend/gql-backend/methods/get-dashboards/get-dashboards.ts
@@ -10,7 +10,7 @@ import {
     InsightsDashboardNode,
     InsightsDashboardsResult,
 } from '../../../../../../../graphql-operations'
-import { ALL_INSIGHTS_DASHBOARD } from '../../../../constants'
+import { ALL_INSIGHTS_DASHBOARD } from '../../../../../constants'
 import {
     InsightDashboard,
     InsightsDashboardOwner,
