+++ client/web/src/enterprise/insights/core/backend/utils/create-line-chart-content.ts
@@ -6,13 +6,11 @@ import { Series } from '@sourcegraph/wildcard'
 
 import { InsightDataSeries, SearchPatternType } from '../../../../../graphql-operations'
 import { PageRoutes } from '../../../../../routes.constants'
-import { DATA_SERIES_COLORS } from '../../../constants'
 import { BackendInsight, InsightFilters, SearchBasedInsightSeries } from '../../types'
 import { BackendInsightDatum, SeriesChartContent } from '../code-insights-backend-types'
 
 import { getParsedSeriesMetadata } from './parse-series-metadata'
 
-export const DATA_SERIES_COLORS_LIST = Object.values(DATA_SERIES_COLORS)
 type SeriesDefinition = Record<string, SearchBasedInsightSeries>
 
 /**
