+++ client/web/src/enterprise/insights/core/index.ts
@@ -1,4 +1,3 @@
-export * from './constants'
 export * from './types'
 
 export * from './backend/code-insights-backend-context'
